refactor(ingress): hoist route.decided + correlation twins to protocol
The two ingress arms may not import each other (openomni↛channels = 0/0),
so their once byte-identical route.decided recorders and wait-correlation
folds were cloned — and drifted. Hoist the PURE parts to protocol so one
core serves both:

- Ingress.routeStreamId / routeDecisionsEquivalent / routeDecidedFact
  (+ ROUTE_DECIDED_FACT_TYPE): the owner-stream key, the replay equivalence
  gate, and the fact-payload builder. Both recorders collapse to (build via
  shared fn) + (append); the append stays per-side. The brain's internal
  recorder now records through the SAME scoped LedgerAppend.port() the router
  uses, replacing raw Storage.get().ledger (ledger-audit finding).
- Wait.waitTierLevels / legacyTierLevels / waitPinsAllowClaim: the pure
  precedence tier-level builders + pin gate. Store reads and candidate
  reduction stay per-side. The drifted PendingAsk externalMessageId fallback
  (channels had it, dispatch lacked it) is converged onto one code path via
  the unified { correlation, externalMessageId } input — inert on the
  dispatch slice (passes no externalMessageId), identical output, no drift.

route.decided strings/stream/equivalence and the wait lookup stay
byte-frozen; p2-ledger-baseline + the equivalence gate stay green.
