+++ packages/channels/src/router/routing-resolution.ts
@@ -375,40 +375,13 @@ function resolveKernelRoute<Event extends Gateway.DeliveredEvent>(
   };
 }
 
-// Route owner-stream key (#510 review fix F1): normalizer-minted inbound ids
-// are only unique WITHIN a channel — telegram normalizer ids are per-chat
-// counters and the github normalizer fallback is
-// `${eventKey}-${issueNumber}-${sender}-${len}` — so the stream identity
-// carries the surface + workspace + channel scope. Without it a colliding id
-// from another channel (or an attacker-chosen channel) could preempt or
-// replay a foreign decision. Each component is URI-encoded (delimiter
-// safety): the protocol schemas allow plain strings, so a ":" inside a
-// channel or id could otherwise forge a foreign scope's key (e.g.
-// channel "C1" + id "x:5" colliding with channel "C1:x" + id "5").
-function routeStreamId(event: Gateway.DeliveredEvent): string {
-  const component = (value: string | undefined) => encodeURIComponent(value ?? "");
-  return `route:${component(event.surface)}:${component(event.workspace)}:${component(event.channel)}:${component(event.id)}`;
-}
-
-// Replay equivalence gate (#510 review fix F2): a cas_conflict means this
-// inbound was ALREADY decided. The recorded decision and the fresh one must
-// agree on every execution-shaping field — stage, outcome, target, sessionId,
-// runId, pendingInteractionId — before the redelivery may proceed. Fields
-// like traceId/time/reason/factsUsed are delivery-local and deliberately
-// excluded.
-function routeDecisionsEquivalent(
-  recorded: Ingress.RoutingDecisionPayload,
-  fresh: Ingress.RoutingDecisionPayload,
-): boolean {
-  return (
-    recorded.stage === fresh.stage &&
-    recorded.outcome === fresh.outcome &&
-    recorded.target === fresh.target &&
-    recorded.sessionId === fresh.sessionId &&
-    recorded.runId === fresh.runId &&
-    recorded.pendingInteractionId === fresh.pendingInteractionId
-  );
-}
+// The route owner-stream key (#510 F1) and the replay equivalence gate
+// (#510 F2) are the PURE parts of this recorder, hoisted to protocol
+// (`Ingress.routeStreamId` / `Ingress.routeDecisionsEquivalent` /
+// `Ingress.routeDecidedFact`) so this external arm and the brain's internal
+// arm cannot drift. Only the append below stays per-side — it holds the
+// router's scoped `LedgerAppend.port()` and throws the channels-local typed
+// error.
 
 // #510 C3 ruling 1 — the routing decision is a decision-class fact on the
 // single-fact owner stream `route:<surface>:<workspace>:<channel>:<id>`
@@ -444,7 +417,7 @@ function recordRouteDecided(
   }
   let appended: ReturnType<typeof ledger.append>;
   try {
-    appended = ledger.append({ streamId, type: "route.decided", data: decision }, 0);
+    appended = ledger.append(Ingress.routeDecidedFact(streamId, decision), 0);
   } catch (error) {
     throw new IngressRoutingError(
       "route_record_failed",
@@ -456,7 +429,7 @@ function recordRouteDecided(
   let recorded: Ingress.RoutingDecisionPayload;
   try {
     const fact = ledger.headFact(streamId);
-    if (fact === undefined || fact.type !== "route.decided") {
+    if (fact === undefined || fact.type !== Ingress.ROUTE_DECIDED_FACT_TYPE) {
       throw new Error(`stream ${streamId} conflicted without a recorded route.decided fact`);
     }
     recorded = Ingress.Events.RoutingDecision.schema.parse(fact.data);
@@ -467,7 +440,7 @@ function recordRouteDecided(
       decision,
     );
   }
-  if (!routeDecisionsEquivalent(recorded, decision)) {
+  if (!Ingress.routeDecisionsEquivalent(recorded, decision)) {
     throw new IngressRoutingError(
       "route_replay_divergent",
       `redelivered inbound diverges from its recorded routing decision: recorded ${recorded.stage}/${recorded.outcome}, fresh ${decision.stage}/${decision.outcome}`,
@@ -490,7 +463,7 @@ export function resolveAndRecordRoute<Event extends Gateway.DeliveredEvent>(
   const decision = Ingress.Events.RoutingDecision.schema.parse(resolution.decision);
   // Redelivery passes the equivalence gate or fails closed — execution below
   // always uses the fresh decision with its own fresh resolution.
-  const effective = recordRouteDecided(routeStreamId(event), decision);
+  const effective = recordRouteDecided(Ingress.routeStreamId(event), decision);
   // Observe-only projection — strictly after the append (or after the gated
   // equivalent replay); lossy by contract, published through the injected
   // sink (channels never imports the observation channel). A divergent
